:sparkles: feat: Add ShippingController and related routes and models
This commit adds a new controller for handling shipping-related
operations, such as viewing shipping details on a map. It also adds
a new relationship method to the Shipment model to access the customer
entity. Additionally, it modifies some existing routes and controllers
to improve code readability and formatting. It also adds some new
methods to the EntityController for viewing the current and future
status of production sites, which require a flood API integration.